LPS-201449 SF, improves packaging, moves all component classes *Manager to the "manager" package keeping in the "util" package those classes *Util which are no component and have static methods
